requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.7.0
google-api-python-client>=2.0.0
tqdm>=4.0.0
pyautogui>=0.9.53  # optional, for GUI mode
